Rename parser_v2 to simply... parser! Also, predict is actually all wired up to get float amount not money type
